// packages/sdk/typechain/ConcentratedLiquidityBasePriceOracle.ts
export declare namespace ConcentratedLiquidityBasePriceOracle {
  export type AssetConfigStruct = {
    poolAddress: string;
    twapWindow: BigNumberish;
    baseToken: string;
  };

  export type AssetConfigStructOutput = [string, BigNumber, string] & {
    poolAddress: string;
    twapWindow: BigNumber;
    baseToken: string;
  };
}